TPCRM2-19: Functionality for Create customer page and Services listing page except localization.

+++ database/migrations/2016_06_27_114508_create_services_table.php
@@ -15,7 +15,7 @@ class CreateServicesTable extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title', 100);
-            $table->integer('hours');
+            $table->decimal('hours', 10, 2);
             $table->decimal('rate', 10, 2);
         });
     }
